LPS-93270 Rename
