+++ modules/apps/headless/headless-delivery/headless-delivery-impl/src/main/java/com/liferay/headless/delivery/internal/jaxrs/exception/mapper/BlogPostingHeadlineExceptionMapper.java
@@ -14,7 +14,7 @@
 
 package com.liferay.headless.delivery.internal.jaxrs.exception.mapper;
 
-import com.liferay.blogs.exception.EntryContentException;
+import com.liferay.blogs.exception.EntryTitleException;
 import com.liferay.portal.kernel.util.StringUtil;
 
 import javax.ws.rs.core.MediaType;
@@ -24,7 +24,7 @@ import javax.ws.rs.ext.ExceptionMapper;
 import org.osgi.service.component.annotations.Component;
 
 /**
- * Converts any {@code EntryContentException} to a {@code 400} error.
+ * Converts any {@code EntryTitleException} to a {@code 400} error.
  *
  * @author Víctor Galán
  * @review
@@ -33,21 +33,21 @@ import org.osgi.service.component.annotations.Component;
 	property = {
 		"osgi.jaxrs.application.select=(osgi.jaxrs.name=Liferay.Headless.Delivery)",
 		"osgi.jaxrs.extension=true",
-		"osgi.jaxrs.name=Liferay.Headless.Delivery.EntryContentExceptionMapper"
+		"osgi.jaxrs.name=Liferay.Headless.Delivery.BlogPostingHeadlineExceptionMapper"
 	},
 	service = ExceptionMapper.class
 )
-public class EntryContentExceptionMapper
-	implements ExceptionMapper<EntryContentException> {
+public class BlogPostingHeadlineExceptionMapper
+	implements ExceptionMapper<EntryTitleException> {
 
 	@Override
-	public Response toResponse(EntryContentException ece) {
+	public Response toResponse(EntryTitleException ete) {
 		return Response.status(
 			400
 		).type(
 			MediaType.TEXT_PLAIN
 		).entity(
-			StringUtil.replace(ece.getMessage(), "Content", "Article body")
+			StringUtil.replace(ete.getMessage(), "Title", "Headline")
 		).build();
 	}
