+++ modules/apps/headless/headless-delivery/headless-delivery-impl/src/main/java/com/liferay/headless/delivery/internal/jaxrs/exception/mapper/KnowledgeBaseArticleArticleBodyExceptionMapper.java
@@ -14,7 +14,8 @@
 
 package com.liferay.headless.delivery.internal.jaxrs.exception.mapper;
 
-import com.liferay.document.library.kernel.exception.FileSizeException;
+import com.liferay.knowledge.base.exception.KBArticleContentException;
+import com.liferay.portal.kernel.util.StringUtil;
 
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
@@ -23,30 +24,30 @@ import javax.ws.rs.ext.ExceptionMapper;
 import org.osgi.service.component.annotations.Component;
 
 /**
- * Converts any {@code FileSizeException} to a {@code 400} error.
+ * Converts any {@code KBArticleContentException} to a {@code 400} error.
  *
- * @author Alejandro Hernández
+ * @author Víctor Galán
  * @review
  */
 @Component(
 	property = {
 		"osgi.jaxrs.application.select=(osgi.jaxrs.name=Liferay.Headless.Delivery)",
 		"osgi.jaxrs.extension=true",
-		"osgi.jaxrs.name=Liferay.Headless.Delivery.FileSizeExceptionMapper"
+		"osgi.jaxrs.name=Liferay.Headless.Delivery.KnowledgeBaseArticleArticleBodyExceptionMapper"
 	},
 	service = ExceptionMapper.class
 )
-public class FileSizeExceptionMapper
-	implements ExceptionMapper<FileSizeException> {
+public class KnowledgeBaseArticleArticleBodyExceptionMapper
+	implements ExceptionMapper<KBArticleContentException> {
 
 	@Override
-	public Response toResponse(FileSizeException fse) {
+	public Response toResponse(KBArticleContentException kbace) {
 		return Response.status(
 			400
 		).type(
 			MediaType.TEXT_PLAIN
 		).entity(
-			fse.getMessage()
+			StringUtil.replace(kbace.getMessage(), "Content", "Article body")
 		).build();
 	}
